+++ web-frontend/modules/core/module.js
@@ -80,12 +80,14 @@ export default defineNuxtModule({
         baserowBuilderDomains: [],
         baserowRowPageSizeLimit: 200,
         baserowUniqueRowValuesSizeLimit: 100,
+        baserowMaxFieldTextLength: 1_000_000,
         baserowDisableSupport: '',
         baserowIntegrationsPeriodicMinuteMin: '1',
         mediaUrl: 'http://localhost:4000/media/',
         sentryDsn: '',
         sentryEnvironment: '',
-        sentryTracesSampleRate: '0.1',
+        sentryTracesSampleRate: '0.01',
+        sentryReplaysOnErrorSampleRate: '0.1',
       }
     )